% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChaosGameTotem2D.R
\name{game2D}
\alias{game2D}
\title{ChaosGameTotem2D}
\usage{
game2D(initial, q, iteration = 20000)
}
\arguments{
\item{initial}{在平面上任意一個初始點及三個(含)以上頂點，為矩陣型態。}

\item{q}{步長比例，設定於[-0.5,0.5]範圍裡，為數值型態。}

\item{iteration}{迭代次數，預設=20000。}
}
\description{
進行Chaos Game，依照使用者在二維所設置的初始點與隨機的頂點及所走的步長，迭代數次後點出一個圖騰。
}
\details{
Chaos Game是依照玩家給予在二維上一個初始點與任意三個(含)以上的頂點及步長比例。利用隨機選擇玩家給予的任一個頂點，並將其與初始點的距離依照步長比例，畫出一個點，此點即為下一個起點。反覆這些步驟，迭代多次即可畫出明顯的圖騰。
}
\examples{
initial = matrix(c(40,20,55,80,1,-1,100,25),nrow = 4,ncol = 2,byrow = TRUE)
#初始點(40,20)、三個頂點(55,80)、(1,-1)、(100,25)
game2D(initial,q=0.5)
# returns 一個三角形圖騰

initial = matrix(c(50,40,52,80,10,50,75,-5,30,0,90,45),nrow = 6,ncol = 2,byrow = TRUE)
#初始點(50,40)、五個頂點(52,80)、(10,50)、(75,-5)、(30,0)、(90,45)
game2D(initial,q=-0.35,iteration = 5000)
# returns 一個迭代5000次的五角圖騰

}
